server/requirements.txt
asgiref==3.3.4
Django==2.2
django-cors-headers==3.7.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysqlclient==2.0.3
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
